src/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
scikit-learn==1.4.2
numpy==1.26.4
pydantic==2.6.4
